Strict boxed: Add Unsafe modules
