+++ vector/vector.cabal
@@ -144,6 +144,8 @@ Library
 
         Data.Vector.Strict.Mutable
         Data.Vector.Strict
+        Data.Vector.Strict.Mutable.Unsafe
+        Data.Vector.Strict.Unsafe
 
         Data.Vector.Mutable
         Data.Vector
